core: move dynamic enum registry into core
Now that the registry keys on an opaque token rather than a config
option, it no longer belongs to the config layer. Relocate it to core
so both config and any other subsystem can use it.
