+++ src/meson.build
@@ -191,7 +191,6 @@ common_sources = [
   'trx/av/image.c',
   'trx/av/video.c',
   'trx/config/common.c',
-  'trx/config/dynamic_enum.c',
   'trx/config/enum.c',
   'trx/config/file.c',
   'trx/config/map.c',
@@ -205,6 +204,7 @@ common_sources = [
   'trx/core/bson/write.c',
   'trx/core/colors.c',
   'trx/core/csv.c',
+  'trx/core/dynamic_enum.c',
   'trx/core/enum_map.c',
   'trx/core/event_manager.c',
   'trx/core/filesystem.c',
