+++ src/trx/config.h
@@ -1,5 +1,4 @@
 #pragma once
 
 #include <trx/config/common.h>
-#include <trx/config/dynamic_enum.h>
 #include <trx/config/vars.h>
