+++ src/trx/game/console/cmd/config.c
@@ -1,8 +1,8 @@
 #include <trx/game/console/cmd/config.h>
 
 #include <trx/config.h>
-#include <trx/config/dynamic_enum.h>
 #include <trx/core/colors.h>
+#include <trx/core/dynamic_enum.h>
 #include <trx/core/enum_map.h>
 #include <trx/core/memory.h>
 #include <trx/core/strings.h>
